일부 int -> INT8로 수정 및 right 보내는거 삭제 (버그 가능성O)

+++ Client/Client/GameFramework.h
@@ -59,7 +59,7 @@ public:
 	void OnProcessingSocketMessage(HWND hWnd, UINT nMessageID, WPARAM wParam, LPARAM lParam);
 	void OnProcessingWindowMessage(HWND hWnd, UINT nMessageID, WPARAM wParam, LPARAM lParam);
 
-	int GetClientIdFromTcpClient() const { return m_pTcpClient->GetClientId(); }
+	INT8 GetClientIdFromTcpClient() const { return m_pTcpClient->GetClientId(); }
 
 	static UCHAR* GetKeysBuffer();
 	static int GetMainClientId() { return m_nMainClientId; }
